app/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
install==1.3.5
Mako==1.3.0
MarkupSafe==2.1.3
paho-mqtt==1.6.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
pydantic-core==2.10.1
pyscopg2==66.0.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
